server/requirements.txt
pycups
pillow
qrcode
requests
python-socketio